+++ internal/meta/extend.go
@@ -48,7 +48,7 @@ var FieldProps = map[string]bool{
 	"mandatoryDependsOn": true, "allowOnSubmit": true, "inListView": true,
 	"inStandardFilter": true, "searchIndex": true, "length": true, "precision": true,
 	"columns": true, "width": true, "gridEditMode": true, "collapsible": true, "bold": true,
-	"gridSort": true, "gridSortable": true, "gridExport": true, "gridSelect": true,
+	"gridSort": true, "gridSortable": true, "gridExport": true, "gridSelect": true, "gridIndex": true,
 	"showFileName": true,
 	"optionColors": true, "options": true, "permlevel": true,
 }